[IMP] l10n_in*:  Support Debit note
Before this commit:
There was no support for the generation of electronic documents (EDI) for Debit
notes. This resulted in confusion and errors when creating Debit notes, as the
system was still generating EDI for invoices. Furthermore, there was no
integration between the "l10n_in_edi_ewaybill" and "account_debit_note" modules,
which prevented the generation of e-way bills for Debit notes.

After this commit:
The system now fully supports the generation of electronic documents (EDI) for
Debit notes. This has eliminated any confusion or errors that were previously
encountered during the creation of Debit notes. In addition, the integration
between the "l10n_in_edi_ewaybill" and "account_debit_note" modules has been
implemented, allowing for the generation of e-way bills for Debit notes.

Task ID = 3262416