[IMP] web: Dropdown unification
We currently have too many different implementation of the same concern:
an input with a dropdown/popover/select menu that can be customized,
display a list of elements, let the user navigate with the keyboard,
can be closed with escape, can be closed with an outside click. Also,
the current navigation systems does not handle inputs properly.

This commit addresses that by doing multiple things:

1. It tries to improve the Dropdown component with a cleaner props API
while keeping its ease of use.

2. The navigation has been improved to support inputs and better customization.

3. The Dropdown now uses the popover/overlay service which will cause
less z-index issues.