[IMP] web_tour: make tour service work with shadow DOM
This commit updates the tour service so that triggers/actions can
be run in a shadow DOM.

In order to configure the tour to use the shadow DOM, the following
configurations options are added:
- `shadowDOM` on the tour configuration. This key must be
a valid selector matching the target shadow host element. All triggers/
actions of the tour will then be performed in its `ShadowRoot`.

- `shadowDOM` on any step of the tour. This value could either be
a valid selector matching the target shadow host element or false.
The behavior is the following:
  - undefined: the step inherit the `shadowDOM` configuration from its
  tour.

  - string: the step triggers/actions will be performed in the
  `ShadowRoot`related to the host matching the given selector.

  - false: the step triggers/actions will be performed in the light
  DOM.