[FIX] hr, hr_attendance: wrong status color, replace with o_status
The text-color in association with fa-circle classes is not displaying
the proper color.

It is because in commit (2705eb056c15dd4535dc525e07979ecaa38f7113)
 we removed the mixin that was handling the exception for the
'fa-circle' class when used with contextual classes to display a status
graphical elements.

This commit replaces the fa-circle class with the `o_status` class which
allows to apply a bg-* color class instead of a text-*.

Note: there are still some fa-circle in association with text-* class
that will have to be adapted but couldn't be with the `o_status` because
it requires a deeper implementation to achieve the same visual result.

task-2818586