[IMP] web: improve input colors
== ISSUE ==

If you go to the Payroll Module, you can see that the checkboxes on the
main dashboard have a different color for the `background`
and the `border`.

This is due to inconsistencies in the color theme.

There was also an issue with the checkboxes in Accounting.
If you go to Accounting > Configuration > Currencies
and hover a checkbox, the checkboxes have a different background than
the one in Payroll which is not consistent.

== After this commit ==

The `background` and `border` color of the inputs are set
to the enterprise one. We also handle this in dark mode.

task-2818586