[FIX] hr_holidays: colors issues
=== ISSUE ===

If you navigate to Time Off > Configuration > Accrual Plans and select
a plan, there is a grey background which extends outside its container.

There is also a readability issue since both (background and kanban item)
are grey.

=== AFTER ===

We fix the margin of the element  with the correct values.

We also slightly modify the colors of elements to improve contrast.

task-3326315
part of task-3326263