[IMP] web: avoid opening record on many2one field in list view
Before this commit, when the `Many2oneField` component is overriden
the list view will mount the new widget (let's say
`CustomMany2oneField`) instead of just using the formatter.
It means the rendering will be different than the generic
component when the record is in readonly. The user can click on
the `display_name` displayed in that `CustomMany2oneField` component
to open record which is not possible when the generic component is
used instead.

This commit adds a new getter `canOpen` in the many2one to integrate
the readonly and edit mode and so `hasExternalButtonLink` getter can
be removed and replaced by that new getter. And it also creates a new
component `ListMany2oneField` used only in the list view to disable
the "canOpen" button when many2one widget is mount in the list view.

Problem found in task-3251672