[FIX] web_editor: wysiwyg: prevent image in iframe crash
Instanciate the wysiwyg on an iframe you have no control on (that is,
you don't know what code is loaded inside). Specifically,
don't have the ifram load bootstrap tooltip.

Insert an image in that iframe with the media dialog.
Click on that image to see the OdooEditor's toolbar

Before this commit, there was a crash, because tooltip was not a function on the target.
That is expected since tooltip was not loaded in the iframe.

After this commit, there is no crash, as we avoid invoking tooltip
if it is not there.