[IMP] web: enhance property field display in kanban view
This commit improves the appearance of property fields in the Kanban view of
item Kanban. The boolean field now has a badge appearance with a label.
Additionally, a label will be shown for the integer, float, date, and datetime
property fields. There are no changes for the remaining fields.

Task-3291930