[IMP] pos: Improve demo data
This PR updates the demo data of the point_of_sale and pos_restaurant.
The new demo data have products more in adequation to what can be sell
in a point of sale. In the data, there is also past sessions with past
orders and open sessions with open orders.