[FIX] base: prevent traceback of XMLSyntaxError and replace Validation with User
when the user did a mistake in any view from UI, like missing the closing tag or
did any syntactical mistake in any view (standard or custom view) during that
time  XMLSyntaxError is generated.

Steps To Produce:- (This is only for example)

1) install 'industry_fsm' and make debug mode on
2) Go to 'Project', Open any Project
3) Go to debug mode > Edit View: Kanban
4) remove the 'closing tag' of the 'attribute tag' and try to save.

The error will be generated.

Note:- To avoid 'Noise' in sentry we changed 'ValidationError' with 'UserError'

Applying these changes will resolve this issue.

sentry - 4157344102