[FIX] stock: fix can not find generic route in stock
when the user deletes the route of the `stock` module that is present in the
data file of`Purchase Stock`. while installing or upgrading the
`MRP Subcontracting`  module the user gets an error as the stock route
referenced from the `Purchase Stock` module has been deleted.

Steps to reproduce :
1. Install the Inventory module.
2. Install the purchase and MRP module.
3. Open the Inventory app.
4. Go to configuration > settings.
5. Enable Multi-steps routes and save them.
6. Open configuration > Routes.
7. Delete the `Buy` route.
8. Install the `MRP Subcontracting` module.
9. Traceback generated.

This commit will raise an UserError while deleting the route which is created
from the data file.

sentry - 4128085959