[FIX] account: prevent trace back while the customer has no address.
If you select two or more invoices from the Tree view, and at least one of the
invoices does not have a customer address, and then click the "Send" button, an
error will occur. This is because the code is trying to get the "False" key from
a dictionary, but the dictionary does not contain that key.'

Steps to Reproduce
- Install `account'  modules.
- Create a new customer without adding the address.
- Open an Invoices module, and Select two or more invoices, In which a minimum
 of one does not have a customer address.
- Click a 'send' in action. An Error will be generated.

Applying this commit will resolve this issue.

sentry-4054584178