[FIX] crm: prevent zero or negative weights to member in lead assignment
"<class 'ValueError'>: "Total of weights must be greater than zero" while
evaluating 'model._cron_assign_leads()'" is generated when cron is trying to
assign sales members to the leads. This is because the weight is zero, but when
we pass the weights in choice, it is always more than zero.

This commit solves the above issue by ensuring that if the weights are zero
or negative, it assigns 1 to the weights.

sentry-4199398621