[FIX] web_editor: toolbar position update on scroll
The floating toolbar position is updated on scroll. For that, the
scroll container is calculated when the Wysiwyg is started.

Before this commit, the scroll container was (sometimes) calculated as
the document element, which is not the correct one.

The scroll container on v15.0 is always '.o_action_manager .o_content'
when the windown is not small (when there's a floating toolbar), which
allows us to simplify things.

task-3263463