[IMP] planning: add default avatar of resource & employee
Description of the issue/feature this PR addresses:
In planning form view, when new resource and employee is created on fly, default
avatar of Initial name should be set instead of placeholder (grey icon).

This commit add default avatar of initial name when new resource and employee is
created. This issue is coming from hr module as avatar of employee is set to a
placeholder (grey icon) when avatar or image is not found. So, we have to change
in hr module which change placeholder to function which give required avatar to
employee when created.

task- 3251689