[FIX] web: display the date/datetime property field in kanban
Before this commit,
In the knowledge app, the date and datetime property fields are not visible in
the item-kanban.

Reason:
In the previous version's date and datetime property field values are of
string type in a non-ISO format, which are then converted into datetime objects.
However, from the version 16.2, the values of date and datetime property fields
are of string type in the ISO format, so it can not be parsed into the datetime
objects.

After this commit,
Now the date and datetime property fields are visible in the item-kanban.

Task-3324573