[FIX] website: redirect only when url_to given
While creating redirects/rewrite if the user does not enter
any url in the 'Url to' field then during redirection, the error
'NotFound: 404 Not Found: The requested URL was not found on the server'
will be produced.

This commit will prevent the traceback and will only redirect when
the url_to is given.

sentry-4206504892