[FIX] l10n_fr_hr_holidays: adapt checks to avoid errors
Before this commit, one test was failing because some specific cases
in the leave creation would lead to an error due to an assert.
This commit changes the way the check is made for a new date_to in
french companies, ensuring that it only occurs if there's only one
calendar and one company involved. A more torough description is set
as commentary of the `_get_fr_date_to` method.

task-3344770