[FIX] base_vat: generate warning while select country as a france and vat fr
When select 'France' into 'country' and enter 'FR' into 'Tax ID'
save the record it will generate trackback.

Steps to produce:
 1. Into Invoice settings select 'verify Vat Number'
 2. Go to customer create new record
 3. Select country as a 'France' and enter Vat 'FR'

Error: A traceback appears: KeyError: ('res.partner',
<function ResPartner._check_vies at 0x7feb8d0a4700>, 'IT13599571000')

Therefore in this case while we select country as 'france' and Vat 'fr',
it will throws an exception,in this commit instead of exception
we used warning to generate as warning message.

Sentry-4209298700