[IMP] payment: payment_status and confirm pages design
The payment status page have been reworked an now displays a maximum of
one entry. Thus it was necessary to adapt the design. The goal is to 
adapt both the payment/status and confirmation pages to ease the visual
transition and increase the coherency between the two.

task-3332066

part of: 3125913