[IMP] payment stripe: migrate to a direct payment flow
The Reserve Bank of India (RBI) issued a directive that introduces
additional security measures for recurring payments on India-issued
cards. Merchants now need to register e-mandates and issue pre-debit
notifications to customers. Stripe has worked with a partner platform to
support that, but we must manipulate the PaymentIntent and SetupIntent
objects directly through their dedicated API, which the Checkout API
does not allow. 

Therefore, we must now integrate with the Elements API and implement a
direct payment flow instead of the current payment with a redirection
flow powered by the Checkout API.

task-3322020