[IMP] sale: add client_order_ref in portal view
Purpose
=======
The customer reference is a string that can be added to a SO
(in Other info page). It is well visible when downloading or printing
a SO but not in the overview in the Portal. 

Spec
=====
Add the Customer Reference in Sales Order and Quotations if it has been filled.

Task-3355199