[IMP] mail,portal,test_{mail_full|website}:add language support for portal link
Before this change, the portal link in the email always redirected to a page in
the default language (for example while sending a quotation). Now, the portal
link redirect to a page in the customer language if it is available.

We add methods in ir_http in portal to easily localize URL so that it can be
used in other modules.

Technical notes: when "website" is installed, the portal language is restrained
to the website languages. So having a partner with a given language doesn't
guarantee that that language is available in the portal. That's why we check
the availability of the language before prepending the language to the URL
(otherwise it leads to a 404). For that, we call "_get_frontend_langs" with
request.is_frontend to True.

Tests have been added in test_mail_full to test the portal behavior alone and
in test_website to test the portal behavior with the website installed. This is
necessary because installing website restraint the available language of portal
to the one of the main website.

In test_mail_full, we patch the method get_available of res_lang to the "base"
behavior. The goal is to ensure that get_available is not limited to the
language of the main website when website module is installed and keep the
behavior of the portal module without website module regarding the language.
Indeed, when installing the module website, some language that were available
in the portal become unavailable if not installed on the main website.

Task-2716028