[IMP] forum: display related forums links in the sidebar
Display related forums links in the sidebar.

Technical note: we add the value "related_forums" in _prepare_user_values
method to display the other forums of the website in the side menu on various
pages. That value is only added when the "forum" value is a forum indicating
that the page is currently displaying one of the forum. In order to do that, we
need to test the type of the "forum" variable as it can have multiple types:
True, None, a forum recordset, ...

Task-3354382