[FIX] clipboard: fix behaviour when selection is updated before the paste
Cut/Copy:
After a cut/copy operation, if the content of cut/copy zone was altered before
paste,  the content was pasted w/o any style. Also, the cut zone was not
destroyed in case of cut operation.

This fix solves the problem by using the cells stored inside the Clipboard
Plugin instead of the current state of the cells in the grid.

Also, this commit restricts pasting the content multiple times after a
cut operation and clears the clipboard after a cut and single paste as well.

Only Format:
Perform 'Paste' or 'Paste - Only Value' on content from outside the sheet.
Now,'Paste - Only Format' would paste the content as well.

This fix solves the problem by handling content in case the pasteOption is
'onlyFormat'.
Task ID : 3135623