[FIX] mass_mailing: traceback on read-only iframe
Since [1] the `value` prop is no longer part of the standard field
props, and its value must be accessed as
`this.props.record.data[this.props.name]`.

This commit fixes a traceback when opening a mailing in read-only mode,
which resulted from reading `length` of `this.props.value`, which is
undefined.

[1]: 688986f888f2fe2371d58b74ded81315ba6bb353