[FIX] mail: decrement unread count value if message is deleted
Before this commit:
In discuss, if user A sends the message to user B, and for user B if chat of
user A is out of focus then there shows a count of 1 unread message.
Now, if user A deleted the message before user B opens the chat then
the value of the unread message count was not decremented even if the message
was deleted.

After this commit:
If the message is deleted then the value of the unread message count is
decremented on refreshing the page.

Task- 2746505