[FIX] project,portal: fix long name issue in portal view of the project
Before this commit, in the task's list view, if the assignee's name was
excessively long, it would overflow beyond the boundaries of the table.
Similarly, within the chatter section, an excessively long name would ruin
the layout and appearance of the section.

In this commit, The name is now shortened by applying the 'text-truncate'
Bootstrap class.

task-3366464