[FIX] base: fix traceback install demo data
When installing demo data manually after creating the database, there is an
error if the model we try to create doesn't have system create rights.

This is because `xml_import` used to get an environment with root privileges,
but now it uses the environment it has been passed.

The fix is install demo_data with root access rights as even models
that don't have system create rights need to be created with demo data.