--temp-- new animations and keysplines ease in/out
This adds two news animation:
- image_mirror_blur
- image_and_shape_zoom_in (could be an option of image_zoom_in animation)

Introduces keySplines for ease-in/out effect
(demo on the image_and_shape_zoom_in)

TBD:
- To create the mouseleave effect -> we can't use the from/to because
keySplines need a list of value, swapping these values should create the
mouseleave effect.

- Maybe we could apply keySplines by default on every animation and
allow the user to disable it in the animation options allowing to choose
between Ease In/out or Linear. -> To set the animation to linear we
could replace the keySplines curve to a linear one `0,0,1,1`

- Use the mouseleave effect if the user clicks on an animated shape
containing a href