[IMP] website: floating dropdown of search bar
Prior to this commit, the dropdown of the search bar which was displayed
in a navbar had a static position when the viewport was below to lg.
This dropdown pushed the navigation down instead of being in front of
it, creating a layout issue in the offcanvas menu.

To fix this, this commit forces the `dropdown_menu` of the search bar to
be in absolute position.

task-3060986