[REF] core: ConfigCleaner(1) dedicated _build_cli
This commit is part of a larger refactor, see associated PR.

Isolate building the CLI in a dedicated method so that it is possible to
reference the `config` singleton instance when instanciating new options

Isolate loading the default option values in a dedicated method so that
it will be easier (later on) to write test cases.

Rename (partially) `config.casts` to `config.options_index` so that it
feels OK to use that datastructure even outside of cast consideration,
the attribute will be fully renamed in a later commit.