[FIX] web_editor: OdooEditor and wysiwyg don't hint inside t-fields
Before this commit, when typing "/" inside a node that had a t-field attrbute
the OdooEditor displayed its hint in order to spawn the Powerbox.

This is not ideal as the content of a t-field is rather irrelevant.

After tis commit, a t-field node behaves like a t-out: there is no hinting inside them