[IMP] web_editor, website: add text highlight effects
The goal of this commit is to replace the "Toggle underline" option [1]
from the website's text editor by a new "Highlight effects" option which
should allow applying some highlights on the selected text and setting
their color & thickness.

For this purpose, the JS code used for text animations will be updated
to be more generic and handle text highlights too. Moreover, The "Text
Highlight" options will be applied by adding a "Highlight SVG" to the
targeted text element (allowing to adjust their `stroke` and
`stroke-width`). [2]

[1]: Also, the text editor "Toggle strikethrough" option will be moved
to the list of highlight effects in the new option.

[2]: This is a default implementation, which can be tweaked following
CSS and design requirements.

Remarks:

- This commit adds the JS implementation and the options code of the
feature. The "design & CSS" part will be implemented by the design team
(see `DESIGN NOTE:` comments).

- The `--text-highlight-width` and `--text-highlight-color` properties
are used to control the highlight effect's thickness and color.

task-3285817