[IMP] hr_holidays: improve reports ux
Before this commit if we go to the Time off and go to reporting in that select
the to employee  there are graph , calendar, pivot views and there are two
default filter active employee , Employee > Type .

And After that when we go to  to type there are list, pivot view  and by default
there are 4 filter to the list view of the to type and in list view there is
button.

Improvements :

To Employee :
Add the List view and make it default.
And make Current year by  default.

To Type :
Add the graph view and make it default.
Made changes in the filter add Time off Allocations in that filter.
Remove that button and use action method to the record.

task-3389174