[FIX] l10n_be: Cash difference gains/losses should go in account 700
According to the guidance of the Commission des Normes Comptables, all
gains and losses arising from the mandatory cash rounding must be
accounted directly in account 700 Ventes et prestations de services.

A negative cash difference should be accounted as a debit on account 700
and a positive cash difference should be accounted as a credit on
account 700.

https://www.cnc-cbn.be/fr/avis/arrondissement-des-paiements-en-euro-0

This commit also prevents the automatic creation of accounts 999001 and
999002 which have no place in the Belgian CoA.