[FIX] web: a modified field must always be sent to the server
Before this commit, when a field of type one2many or many2many, was
added and afterward deleted, on the onchange or write call to the server
the field wasn't added as a change (because the add and delete command
cancel each other). This can cause issues because when adding the x2many
field, the onchange on the server can have side effects that we want to
remove when deleting the x2many.

Now, we follow the following rule : If a field has been modified, its
value must always be sent to the server for onchange and write.

This commit also avoid to call update on the parent record if the chil
record didn't have changes.

Task 3179751