[IMP] {website_}event: merge phone and mobile fields on registration
Merge the mobile field into the phone field for event registration,
booth and booth registration.

The reasoning is that we ask only for the phone in the attendee form.
And so is the most used for registration. The mobile field is then a
little redundant. Also the values are the same and does not add anything
meaningful.

task-3366899