[FIX] web: fix unused displayName props
This commit ensures that the displayName variable of the calendar
controller is properly used to update the displayName of the view when
the CalendarRenderer is rendered. This fixes an issue in project where
the extension of the displayName in the project calendar controller
( - Tasks by Deadline) would not be shown in the view.

opw-3410987