[IMP] stock,mrp: Forecast by product configuration
Currently it's not possible to have a replenishment based on different
configuration (e.g. different routes, different supplier, BoM,...)

It would be great to see the different options in replenish report.
This configuration is block by a constraint ensuring that the
product, location, company keys are always different.

And a bug that doesn't modify the forecast date of the orderpoint base
on its BoM (it does for supplier in buy case)