[FIX] sale_project,sale_timesheet: fix sale_line_id behavior
Steps to reproduce:
1) Install the sale_timesheet module from the apps.
2) Open any task that has a timesheet associated with it.
3) Remove the partner from the task and try to select the SOL of another partner.
4) Notice that the SOL cannot be selected.

Cause:
-When a timesheet is associated with an SOL and we try to set a different SOL
with a different customer, the system will search for SOLs with that customer
and remaining hours greater than zero.
-In our specific situation, when we try to set the SOL for
YourCompany, Joel Willis, does not have any remaining hours. so, the SOL
is set to false.
-It is not appropriate to change the SOL for a timesheet that is already
associated with a specific SOL if that SOL has remaining hours equal to zero.
In such cases, the timesheet entries lose their relevance.

Fix:
-The sale_line_id will be shown only if the task has a partner_id associated with it.
-The sale_line_id also contains the SOL (Sales Order Line) of the partner's child.

task-3366464